test(io): split shm IO tests by format
